to 二階堂さん(BaseButton削除)
to 二階堂さん
BaseButtonは今使っていないみたいなので、LoginButtonまでうつしました。
基本、オブジェクト志向の概念に則って今使わない実装は削除お願いします。
